Migrate to declarative clap
Mostly, there still needs to be a procedural bit to set the long about
and the default delimiters. The latter could be handled via
dynamically scoped data (by using a default callback, which'd retrieve
a value set by the caller), but it still simplifies the code a lot.

Changes performed at the same time:

- removed `--rs` and `--fs` because they didn't work, and had never
  worked, not sure why I thought they did
- add proper help for the formula modes, this only works in long help
  mode, but it's better formatted than the old ad-hoc version
- convert a few `write!` to implicit params, as I was already using
  that in a few places
- remove unnecessary `use` statements in the binaries, probably
  leftover from the edition 2021 conversion

Bump version as the format of the help has changed a bit.